requirements.txt
# >=1.45: Chromium compatível com Ubuntu 24.04 (Noble); 1.42 pedia libicu70 (só no 22.04)
playwright>=1.50.0,<2
pyyaml==6.0.1
pandas==2.2.1
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic>=2.0,<3
pytest>=8.0
pytest-asyncio>=0.23
